Can now backtrack and change a form selection.
Form Builder saves its entire state in:
$form_state['storage']['objective_form_storage']

This now gets unset when the previous button is pressed, allowing for other
forms to be selected. Also we unset the saved values as a different form
can be presented to the user than the original so retaining the old values
doesn't make much sense.