ISLANDORA-1087: fixes WSOD
https://jira.duraspace.org/browse/ISLANDORA-1087

Fixes orphan DOMNodes associated to deleted Form elements when
processing the XML document on submit. Since all objects used by
XMLForms are stored inside $form_state and serialized by drupal to keep
multistep states,  when the XML document is processed, and after delete
actions are run, we end with DOMNodes stored in the internal registry
that do not have parents, and also DOMAttrib associated to them. When
serialisation of this particular object happens every DOMNode is
converted to an XPath. On unserialise, those XPaths are not valid
agains the XML document and results in an silent exception when the
unserialization process is nested, destroying $form_state and by this
making form steps impossible.