Fix database schema
Drupal expects the database schema to have a key 'not null' and we were
using 'not NULL'. This is not taken into account, so the field defaulted
to allowing NULL. This causes problems with MySQL 5.7 where its an error
to create a primary key that allows null.

This update defines a update hook that fixes existing DB and fixes the
schema to use the correct case.